fix: implement floating points grades with update on the grade form input to use a slider

+++ apps/client/lib/flutter_gen/gen_l10n/app_localizations.dart
@@ -978,7 +978,7 @@ abstract class AppLocalizations {
   ///
   /// In en, this message translates to:
   /// **'Original rating: {rating} stars'**
-  String originalRating(int rating);
+  String originalRating(double rating);
 
   /// Indicator that there are unsaved changes
   ///
